Created a lambda function to re-index the promotions table and provide the necessary data for the pipeline to collect.
Updated the cloud formation script to create the lambda function, it's role and its event hooks to the dynamo streams.

Created a build script and riffraff package for the lambda function.

Ensure cloud formation script references lambda code zip from same place as riffraff puts it!

Fall back to UAT tables if the lambda function name does not end with PROD.

Call the UAT Lambda function -QA so that RiffRaff can deploy it.