Fix (sometimes) broken date picker
After the Node + NPM lib upgrades in #255 we found that the date picker
was sometimes breaking. After lots of debugging/trial and error we
discovered that the datepicker was only broken when the day potion of
the date was > 12. This lead us to believe this was a date locale
problem. As an experiment we tried removing the formatDate change we
were making in DateConfig and this got things working. We aren't sure
why! It doesn't seem to have broken anything else.