Really fix the date picker!
Previously we'd removed the formatDate function. This got the date
picker working, but it meant we were incorrectly formatting a selected
date in the input field. Instead the right approach is to add the
formatting back in and to add a parseDate function, which knows how to
go from our formatted date back to a Date object.