Fix (sometimes) broken date picker (#261)
After the Node + NPM lib upgrades in #255 we found that the date picker was
sometimes breaking. After lots of debugging/trial and error we discovered that
the datepicker was only broken when the day potion of the date was > 12. This
lead us to believe this was a date locale problem.

To fix this we've added a `parseDate` function to the `mdDateLocaleProvider` to
get back to a `Date` object from the UK formatted dates we display.