Upgrade lambdas to Node 22 (#615)
Upgrade the lambdas in this repo to run on Node 22 (previously Node 18, which
will soon stop being supported). This is the latest runtime supported by AWS
currently.

I've tested the changes in CODE. I did this by manually updating the runtime to
NodeJS 22 for the two lambdas. I did this rather than deploy the Cloudformation
changes in the usual way because in Riff Raff/the Cloudformation there's only
one stage - PROD, which the CODE and PROD lambdas belong to.

MembershipSub-Promotions-to-PromoCode-View-Lambda:

This is the lambda which is triggered by events from another Dynamo table. I
made a minor update directly to data in the source table and saw the event get
handled correctly.

MembershipSub-PromoCode-View-Dynamo-to-Salesforce:

This is the lambda which runs on a schedule and upserts data from the View
Dynamo table to Salesforce. I invoked this lambda manually and saw it run
successfully. There is a small discrepancy between the count of records in
Salesforce and in the Dynamo table, but I see this with the Node 18 runtime
too. I don't think it's related to this PR.